requirements.txt
torch
torchvision
opencv-python
pillow
numpy
pandas
tqdm
huggingface_hub
rfdetr
supervision
scipy
pyyaml
inference
